move conf class to new package
